+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix that the test run finished prematurely when the first scenario finished.
 - Log `world.attach` messages as test output.
 - Fix that the number of seconds the test step took were omitted - only the nanoseconds were counted.
+- Start the test items one by one instead of all at once. Instead, queue all test items but only start the scenario and step that started.
 
 ## [0.7.3]