+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this extension will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.5]
+
+- Fix that all tests should not error when cucumber-js reported anything to `stderr`.
+
 ## [0.7.4]
 
 - Skip the `node_modules` when debugging, excluding `cucumber.js`.